requirements.txt
pandas==1.4.2
PyYAML==6.0
requests==2.26.0
